services/moderation_image_catornotclassifier/pyproject.toml
[project]
name = "moderation-image-catornotclassifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.16",
    "faststream[cli,rabbit]>=0.6.4",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
